+++ src/main/java/woowacourse/auth/ui/dto/SignInRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto;
+package woowacourse.auth.ui.dto;
 
 public class SignInRequest {
