+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -8,7 +8,8 @@ import woowacourse.shoppingcart.dao.CartItemDao;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Cart;
-import woowacourse.shoppingcart.entity.ProductEntity;
+import woowacourse.shoppingcart.dao.entity.ProductEntity;
+import woowacourse.shoppingcart.exception.InvalidCustomerException;
 import woowacourse.shoppingcart.exception.InvalidProductException;
 import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
 
@@ -32,10 +33,15 @@ public class CartService {
         final List<Cart> carts = new ArrayList<>();
         for (final Long cartId : cartIds) {
             final Long productId = cartItemDao.findProductIdById(cartId);
-            final ProductEntity productEntity = productDao.findProductById(productId);
+            final ProductEntity productEntity = getProductEntity(productId);
             carts.add(new Cart(cartId, productEntity.toProduct()));
         }
         return carts;
+    }
+
+    private ProductEntity getProductEntity(Long productId) {
+        return productDao.findProductById(productId)
+                .orElseThrow(InvalidCustomerException::new);
     }
 
     private List<Long> findCartIdsByCustomerName(final String customerName) {
