+++ src/main/java/woowacourse/shoppingcart/application/OrderService.java
@@ -12,7 +12,7 @@ import woowacourse.shoppingcart.dao.OrdersDetailDao;
 import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.domain.Orders;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.OrderRequest;
+import woowacourse.shoppingcart.ui.dto.OrderRequest;
 import woowacourse.shoppingcart.exception.InvalidOrderException;
 
 @Service
