+++ src/main/java/woowacourse/shoppingcart/application/dto/PhoneNumberResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.application.dto;
 
 import woowacourse.shoppingcart.domain.customer.vo.PhoneNumber;
