+++ src/main/java/woowacourse/shoppingcart/application/dto/ProductResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.application.dto;
 
 import woowacourse.shoppingcart.domain.Product;
