+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -13,7 +13,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import woowacourse.shoppingcart.entity.CustomerEntity;
+import woowacourse.shoppingcart.dao.entity.CustomerEntity;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 
 @Repository
