+++ src/main/java/woowacourse/shoppingcart/dao/entity/CustomerEntity.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.entity;
+package woowacourse.shoppingcart.dao.entity;
 
 import woowacourse.shoppingcart.domain.customer.Customer;
 import woowacourse.shoppingcart.domain.customer.vo.Account;
