+++ src/main/java/woowacourse/shoppingcart/ui/CartItemController.java
@@ -15,7 +15,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.application.CartService;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.Request;
+import woowacourse.shoppingcart.ui.dto.Request;
 
 @RestController
 @RequestMapping("/api/customers/{customerName}/carts")
