+++ src/main/java/woowacourse/shoppingcart/ui/dto/OrderRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.ui.dto;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
