+++ src/main/java/woowacourse/shoppingcart/ui/dto/PasswordRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.ui.dto;
 
 public class PasswordRequest {
