+++ src/main/java/woowacourse/shoppingcart/ui/dto/PhoneNumberRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.ui.dto;
 
 import woowacourse.shoppingcart.domain.customer.vo.PhoneNumber;
