+++ src/main/java/woowacourse/shoppingcart/ui/dto/Request.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.ui.dto;
 
 public interface Request {
     interface id {
