+++ src/test/java/woowacourse/shoppingcart/dao/CustomerDaoTest.java
@@ -12,7 +12,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
-import woowacourse.shoppingcart.entity.CustomerEntity;
+import woowacourse.shoppingcart.dao.entity.CustomerEntity;
 
 @JdbcTest
 @AutoConfigureTestDatabase(replace = Replace.NONE)
