+++ src/test/java/woowacourse/shoppingcart/ui/OrderControllerTest.java
@@ -26,7 +26,7 @@ import woowacourse.shoppingcart.application.OrderService;
 import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.domain.Orders;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.dto.OrderRequest;
+import woowacourse.shoppingcart.ui.dto.OrderRequest;
 
 @SpringBootTest
 @AutoConfigureMockMvc
